requirements.txt
bcrypt==3.1.7
cassandra-driver==3.20.2
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
pycparser==2.19
six==1.13.0
SQLAlchemy==1.3.12
torch==1.3.1
torchvision==0.4.2
Werkzeug==0.16.0
